Making serializePayload internal + adding test
