+++ bugsnag-plugin-android-okhttp/src/main/java/com/bugsnag/android/okhttp/NoopLogger.kt
@@ -1,0 +1,5 @@
+package com.bugsnag.android.okhttp
+
+import com.bugsnag.android.Logger
+
+internal object NoopLogger : Logger
